requirements.txt
openai
pytest
python-dotenv

